+++ glean-core/rlb/src/private/event.rs
@@ -72,7 +72,9 @@ impl<K: traits::ExtraKeys> EventMetric<K> {
 }
 
 #[inherent]
-impl<K> TestGetValue<Vec<RecordedEvent>> for EventMetric<K> {
+impl<K> TestGetValue for EventMetric<K> {
+    type Output = Vec<RecordedEvent>;
+
     pub fn test_get_value(&self, ping_name: Option<String>) -> Option<Vec<RecordedEvent>> {
         self.inner.test_get_value(ping_name)
     }